Added shears<->compass Multitool, and improved Multitool triggers (#3287)

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/itemstats/enchantments/Abyssal.java
@@ -41,7 +41,7 @@ public class Abyssal implements Enchantment {
 		if (event.getType() == DamageEvent.DamageType.TRUE) {
 			return;
 		}
-		double level = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.ABYSSAL);
+		int level = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.ABYSSAL);
 		if (EntityUtils.isInWater(enemy) || EntityUtils.isInWater(player)) {
 			double multiplier = 1 + CharmManager.calculateFlatAndPercentValue(player, CHARM_DAMAGE, DAMAGE_BONUS_PER_LEVEL * level);
 			event.setDamage(event.getDamage() * multiplier);
