+++ plugins/paper/src/main/java/com/playmonumenta/plugins/itemstats/enchantments/CurseOfAnemia.java
@@ -21,8 +21,8 @@ public class CurseOfAnemia implements Enchantment {
 
 	@Override
 	public void onRegain(Plugin plugin, Player player, double value, EntityRegainHealthEvent event) {
-		double sustenanceLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.SUSTENANCE);
-		double anemiaLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.CURSE_OF_ANEMIA);
+		int sustenanceLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.SUSTENANCE);
+		int anemiaLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.CURSE_OF_ANEMIA);
 		double mult = Sustenance.getHealingMultiplier(sustenanceLevel, anemiaLevel);
 		// If the player has both Anemia and Sustenance, only one enchant will run to boost/reduce depending on the higher level.
 		if (mult < 1) {
