+++ plugins/paper/src/main/java/com/playmonumenta/plugins/itemstats/enchantments/Sustenance.java
@@ -20,8 +20,8 @@ public class Sustenance implements Enchantment {
 
 	@Override
 	public void onRegain(Plugin plugin, Player player, double value, EntityRegainHealthEvent event) {
-		double sustenanceLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.SUSTENANCE);
-		double anemiaLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.CURSE_OF_ANEMIA);
+		int sustenanceLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.SUSTENANCE);
+		int anemiaLevel = plugin.mItemStatManager.getEnchantmentLevel(player, EnchantmentType.CURSE_OF_ANEMIA);
 		double mult = getHealingMultiplier(sustenanceLevel, anemiaLevel);
 		// If the player has both Anemia and Sustenance, only one enchant will run to boost/reduce depending on the higher level.
 		if (mult > 1) {
